requirements.txt
# Green Wave++ — Python dependencies
# Install: pip install -r requirements.txt
#
# For GPU-accelerated inference and training, install the CUDA build of PyTorch:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# CPU-only fall-through works automatically for demos and CI.

# ── Deep learning framework ──────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── Object detection ─────────────────────────────────────────────────────────
ultralytics>=8.1.0          # Ultralytics YOLOv8/v11

# ── Audio processing ─────────────────────────────────────────────────────────
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.26.0
scipy>=1.12.0

# ── Computer vision ──────────────────────────────────────────────────────────
opencv-python>=4.9.0
Pillow>=10.2.0

# ── Training utilities ───────────────────────────────────────────────────────
scikit-learn>=1.4.0
tqdm>=4.66.0
matplotlib>=3.8.0

# ── API server ───────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0

# ── Configuration ────────────────────────────────────────────────────────────
pyyaml>=6.0.1

# ── Optional: SUMO traffic simulation ────────────────────────────────────────
# Install separately via: pip install traci
# Or install with SUMO from https://sumo.dlr.de/
# traci>=1.18.0
